rezerwacjaSal/requirements.txt
django
django-heroku
django-crispy-forms
gunicorn
psycopg2
pytz
whitenoise
dj-database-url
djangorestframework
django-cors-headers
django-rest-knox